Started
renamed the pages, put the information in them. Linked stylesheets to index. Made the body a flex-box container and set the the section and article as a flex-item. when the website is viewed on browser the image will take up it's own row, and put the nav on the bottom before the footer. For non-mobile screen sizes it will display the nav list horizontal using a flex-box and flex-items, then changed the classes width.