Finished just need comments
changed the section ID in the style sheet from sheets to sheet, then set the width of the images. Then hid the nav, footer and article elements when the website is printed. Then set the width of the section#sheet, size1, size2, size3, size4 to inches so it looks better when printed.